+++ src/crypt/ciphers.rs
@@ -1,18 +1,15 @@
-use aes::Aes256;
 #[cfg(feature = "save_kdbx4")]
 use cipher::BlockEncryptMut;
-use cipher::{block_padding::Pkcs7, generic_array::GenericArray, BlockDecryptMut};
-use salsa20::{
-    cipher::{KeyIvInit, StreamCipher},
-    Salsa20,
+use cipher::{
+    block_padding::{Pkcs7, UnpadError},
+    generic_array::GenericArray,
+    BlockDecryptMut, InvalidLength, KeyIvInit, StreamCipher,
 };
-
-use crate::crypt::CryptographyError;
 
 pub(crate) trait Cipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError>;
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError>;
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8>;
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError>;
 
     #[cfg(feature = "save_kdbx4")]
     /// The number of bytes expected by the cipher as an initialization vector.
@@ -27,38 +24,34 @@ pub(crate) trait Cipher {
         Self: Sized;
 }
 
-#[cfg(feature = "save_kdbx4")]
-type Aes256CbcEncryptor = cbc::Encryptor<Aes256>;
-type Aes256CbcDecryptor = cbc::Decryptor<Aes256>;
 pub(crate) struct AES256Cipher {
-    key: Vec<u8>,
-    iv: Vec<u8>,
+    #[cfg(feature = "save_kdbx4")]
+    encryptor: cbc::Encryptor<aes::Aes256>,
+    decryptor: cbc::Decryptor<aes::Aes256>,
 }
 
 impl AES256Cipher {
-    pub(crate) fn new(key: &[u8], iv: &[u8]) -> Result<Self, CryptographyError> {
+    pub(crate) fn new(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> {
         Ok(AES256Cipher {
-            key: Vec::from(key),
-            iv: Vec::from(iv),
+            #[cfg(feature = "save_kdbx4")]
+            encryptor: cipher::KeyIvInit::new_from_slices(key, iv)?,
+            decryptor: cipher::KeyIvInit::new_from_slices(key, iv)?,
         })
     }
 }
 
 impl Cipher for AES256Cipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
-        let cipher = Aes256CbcEncryptor::new_from_slices(&self.key, &self.iv)?;
-
-        let ciphertext = cipher.encrypt_padded_vec_mut::<Pkcs7>(plaintext);
-
-        Ok(ciphertext)
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8> {
+        self.encryptor.clone().encrypt_padded_vec_mut::<Pkcs7>(plaintext)
     }
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError> {
         let mut out = vec![0; ciphertext.len()];
 
-        let cipher = Aes256CbcDecryptor::new_from_slices(&self.key[..], &self.iv[..])?;
-
-        let len = cipher
+        let len = self
+            .decryptor
+            .clone()
             .decrypt_padded_b2b_mut::<Pkcs7>(ciphertext, &mut out)?
             .len();
 
@@ -78,38 +71,31 @@ impl Cipher for AES256Cipher {
     }
 }
 
-#[cfg(feature = "save_kdbx4")]
-type TwofishCbcEncryptor = cbc::Encryptor<twofish::Twofish>;
-type TwofishCbcDecryptor = cbc::Decryptor<twofish::Twofish>;
 pub(crate) struct TwofishCipher {
-    key: Vec<u8>,
-    iv: Vec<u8>,
+    #[cfg(feature = "save_kdbx4")]
+    encryptor: cbc::Encryptor<twofish::Twofish>,
+    decryptor: cbc::Decryptor<twofish::Twofish>,
 }
 
 impl TwofishCipher {
-    pub(crate) fn new(key: &[u8], iv: &[u8]) -> Result<Self, CryptographyError> {
+    pub(crate) fn new(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> {
         Ok(TwofishCipher {
-            key: Vec::from(key),
-            iv: Vec::from(iv),
+            #[cfg(feature = "save_kdbx4")]
+            encryptor: KeyIvInit::new_from_slices(key, iv)?,
+            decryptor: KeyIvInit::new_from_slices(key, iv)?,
         })
     }
 }
 
 impl Cipher for TwofishCipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
-        let cipher = TwofishCbcEncryptor::new_from_slices(&self.key, &self.iv)?;
-
-        let ciphertext = cipher.encrypt_padded_vec_mut::<twofish::cipher::block_padding::Pkcs7>(plaintext);
-
-        Ok(ciphertext)
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8> {
+        self.encryptor.clone().encrypt_padded_vec_mut::<Pkcs7>(plaintext)
     }
 
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
-        let cipher = TwofishCbcDecryptor::new_from_slices(&self.key, &self.iv)?;
-
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError> {
         let mut buf = ciphertext.to_vec();
-        cipher.decrypt_padded_mut::<twofish::cipher::block_padding::Pkcs7>(&mut buf)?;
+        self.decryptor.clone().decrypt_padded_mut::<Pkcs7>(&mut buf)?;
         Ok(buf)
     }
 
@@ -129,24 +115,25 @@ pub(crate) struct Salsa20Cipher {
 }
 
 impl Salsa20Cipher {
-    pub(crate) fn new(key: &[u8]) -> Result<Self, CryptographyError> {
+    pub(crate) fn new(key: &[u8]) -> Result<Self, InvalidLength> {
         let key = GenericArray::from_slice(key);
         let iv = GenericArray::from([0xE8, 0x30, 0x09, 0x4B, 0x97, 0x20, 0x5D, 0x2A]);
 
         Ok(Salsa20Cipher {
-            cipher: Salsa20::new(key, &iv),
+            cipher: KeyIvInit::new(key, &iv),
         })
     }
 }
 
 impl Cipher for Salsa20Cipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8> {
         let mut buffer = Vec::from(plaintext);
         self.cipher.apply_keystream(&mut buffer);
-        Ok(buffer)
+        buffer
     }
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError> {
         let mut buffer = Vec::from(ciphertext);
         self.cipher.apply_keystream(&mut buffer);
         Ok(buffer)
@@ -170,9 +157,8 @@ pub(crate) struct ChaCha20Cipher {
 
 impl ChaCha20Cipher {
     /// Create as an inner cipher by splitting up a SHA512 hash
-    pub(crate) fn new(key: &[u8]) -> Result<Self, CryptographyError> {
-        let iv = crate::crypt::calculate_sha512(&[key])?;
-
+    pub(crate) fn new(key: &[u8]) -> Result<Self, InvalidLength> {
+        let iv = crate::crypt::calculate_sha512(&[key]);
         let key = GenericArray::from_slice(&iv[0..32]);
         let nonce = GenericArray::from_slice(&iv[32..44]);
 
@@ -182,7 +168,7 @@ impl ChaCha20Cipher {
     }
 
     /// Create as an outer cipher by separately-specified key and iv
-    pub(crate) fn new_key_iv(key: &[u8], iv: &[u8]) -> Result<Self, CryptographyError> {
+    pub(crate) fn new_key_iv(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> {
         Ok(ChaCha20Cipher {
             cipher: chacha20::ChaCha20::new_from_slices(key, iv)?,
         })
@@ -191,12 +177,12 @@ impl ChaCha20Cipher {
 
 impl Cipher for ChaCha20Cipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8> {
         let mut buffer = Vec::from(plaintext);
         self.cipher.apply_keystream(&mut buffer);
-        Ok(buffer)
+        buffer
     }
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError> {
         let mut buffer = Vec::from(ciphertext);
         self.cipher.apply_keystream(&mut buffer);
         Ok(buffer)
@@ -215,16 +201,16 @@ impl Cipher for ChaCha20Cipher {
 
 pub(crate) struct PlainCipher;
 impl PlainCipher {
-    pub(crate) fn new(_: &[u8]) -> Result<Self, CryptographyError> {
+    pub(crate) fn new(_: &[u8]) -> Result<Self, InvalidLength> {
         Ok(PlainCipher)
     }
 }
 impl Cipher for PlainCipher {
     #[cfg(feature = "save_kdbx4")]
-    fn encrypt(&mut self, plaintext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
-        Ok(Vec::from(plaintext))
+    fn encrypt(&mut self, plaintext: &[u8]) -> Vec<u8> {
+        Vec::from(plaintext)
     }
-    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptographyError> {
+    fn decrypt(&mut self, ciphertext: &[u8]) -> Result<Vec<u8>, UnpadError> {
         Ok(Vec::from(ciphertext))
     }
