+++ src/crypt/mod.rs
@@ -1,8 +1,11 @@
 #[cfg(feature = "challenge_response")]
 use cipher::generic_array::typenum::U20;
-use cipher::generic_array::{
-    typenum::{U32, U64},
-    GenericArray,
+use cipher::{
+    generic_array::{
+        typenum::{U32, U64},
+        GenericArray,
+    },
+    InvalidLength,
 };
 
 use hmac::{Hmac, Mac};
@@ -10,15 +13,10 @@ use hmac::{Hmac, Mac};
 use sha1::Sha1;
 use sha2::{Digest, Sha256, Sha512};
 
-use crate::error::CryptographyError;
-
 pub(crate) mod ciphers;
 pub(crate) mod kdf;
 
-pub(crate) fn calculate_hmac(
-    elements: &[&[u8]],
-    key: &[u8],
-) -> Result<GenericArray<u8, U32>, CryptographyError> {
+pub(crate) fn calculate_hmac(elements: &[&[u8]], key: &[u8]) -> Result<GenericArray<u8, U32>, InvalidLength> {
     type HmacSha256 = Hmac<Sha256>;
     let mut mac = HmacSha256::new_from_slice(key)?;
 
@@ -34,7 +32,7 @@ pub(crate) fn calculate_hmac(
 pub(crate) fn calculate_hmac_sha1(
     elements: &[&[u8]],
     key: &[u8],
-) -> Result<GenericArray<u8, U20>, CryptographyError> {
+) -> Result<GenericArray<u8, U20>, InvalidLength> {
     type HmacSha1 = Hmac<Sha1>;
     let mut mac = HmacSha1::new_from_slice(key)?;
 
@@ -46,22 +44,22 @@ pub(crate) fn calculate_hmac_sha1(
     Ok(result.into_bytes())
 }
 
-pub(crate) fn calculate_sha256(elements: &[&[u8]]) -> Result<GenericArray<u8, U32>, CryptographyError> {
+pub(crate) fn calculate_sha256(elements: &[&[u8]]) -> GenericArray<u8, U32> {
     let mut digest = Sha256::new();
 
     for element in elements {
         digest.update(element);
     }
 
-    Ok(digest.finalize())
+    digest.finalize()
 }
 
-pub(crate) fn calculate_sha512(elements: &[&[u8]]) -> Result<GenericArray<u8, U64>, CryptographyError> {
+pub(crate) fn calculate_sha512(elements: &[&[u8]]) -> GenericArray<u8, U64> {
     let mut digest = Sha512::new();
 
     for element in elements {
         digest.update(element);
     }
 
-    Ok(digest.finalize())
+    digest.finalize()
 }
