+++ src/db/save.rs
@@ -1,0 +1,18 @@
+impl Database {
+    /// Save a database to a std::io::Write
+    pub fn save(
+        &self,
+        destination: &mut dyn std::io::Write,
+        key: DatabaseKey,
+    ) -> Result<(), crate::error::DatabaseSaveError> {
+        use crate::error::DatabaseSaveError;
+        use crate::format::kdbx4::dump_kdbx4;
+
+        match self.config.version {
+            DatabaseVersion::KDB(_) => Err(DatabaseSaveError::UnsupportedVersion),
+            DatabaseVersion::KDB2(_) => Err(DatabaseSaveError::UnsupportedVersion),
+            DatabaseVersion::KDB3(_) => Err(DatabaseSaveError::UnsupportedVersion),
+            DatabaseVersion::KDB4(_) => dump_kdbx4(self, &key, destination),
+        }
+    }
+}
