+++ src/db/types/custom_data.rs
@@ -1,0 +1,11 @@
+use chrono::NaiveDateTime;
+
+use crate::db::Value;
+
+/// Custom data field for an entry or metadata for internal use
+#[derive(Debug, Default, PartialEq, Eq, Clone)]
+#[cfg_attr(feature = "serialization", derive(serde::Serialize))]
+pub struct CustomDataItem {
+    pub value: Option<Value>,
+    pub last_modification_time: Option<NaiveDateTime>,
+}
