+++ src/db/types/meta.rs
@@ -1,7 +1,9 @@
+use std::collections::{HashMap, HashSet};
+
 use chrono::NaiveDateTime;
 use uuid::Uuid;
 
-use crate::db::{Color, CustomData};
+use crate::db::{BinaryAttachmentId, Color, CustomDataItem};
 
 /// Database metadata
 #[derive(Debug, Default, Eq, PartialEq, Clone)]
@@ -44,9 +46,6 @@ pub struct Meta {
     /// memory protection settings
     pub memory_protection: Option<MemoryProtection>,
 
-    /// custom icons
-    pub custom_icons: CustomIcons,
-
     /// whether the recycle bin is enabled
     pub recyclebin_enabled: Option<bool>,
 
@@ -78,10 +77,10 @@ pub struct Meta {
     pub settings_changed: Option<NaiveDateTime>,
 
     /// Binary attachments in the Metadata header
-    pub binaries: BinaryAttachments,
+    pub binaries: HashSet<BinaryAttachmentId>,
 
     /// Additional custom data fields
-    pub custom_data: CustomData,
+    pub custom_data: HashMap<String, CustomDataItem>,
 }
 
 /// Database memory protection settings
@@ -114,38 +113,4 @@ impl Default for MemoryProtection {
             protect_notes: false,
         }
     }
-}
-
-/// Collection of custom icons
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "serialization", derive(serde::Serialize))]
-pub struct CustomIcons {
-    pub icons: Vec<Icon>,
-}
-
-/// A custom icon
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "serialization", derive(serde::Serialize))]
-pub struct Icon {
-    /// UUID, to reference the icon
-    pub uuid: Uuid,
-
-    /// Image data
-    pub data: Vec<u8>,
-}
-
-/// Collection of binary attachments in the metadata of an XML database
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "serialization", derive(serde::Serialize))]
-pub struct BinaryAttachments {
-    pub binaries: Vec<BinaryAttachment>,
-}
-
-/// Binary attachment in the metadata of a XML database
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "serialization", derive(serde::Serialize))]
-pub struct BinaryAttachment {
-    pub identifier: Option<String>,
-    pub compressed: bool,
-    pub content: Vec<u8>,
 }
